This is now possible:

```python
def test_multiple(testing):
    # Test a model that has multiple inputs and outputs
    model = ell.model.Model()

    layout = ell.model.PortMemoryLayout([int(10)])
    input1Node = model.AddInput(layout, ell.nodes.PortType.real)
    input2Node = model.AddInput(layout, ell.nodes.PortType.real)
    addNode = model.AddBinaryOperation(input1Node, input2Node, ell.nodes.BinaryOperationType.add)
    diffNode = model.AddBinaryOperation(input1Node, input2Node, ell.nodes.BinaryOperationType.subtract)

    output1Node = model.AddOutput(layout, addNode)
    output2Node = model.AddOutput(layout, diffNode)

    map = ell.model.Map(model, ell.nodes.InputNodeList([input1Node, input2Node]),
                        ell.nodes.OutputNodeList([output1Node, output2Node]))

    x = np.array(range(10))
    y = np.array(range(10)) * 2
    i1 = ell.math.DoubleVector(x)
    i2 = ell.math.DoubleVector(y)
    o1 = ell.math.DoubleVector(10)
    o2 = ell.math.DoubleVector(10)
    map.ComputeMultiple([i1, i2], [o1, o2])
    testing.ProcessTest("Testing multiple inputs and outputs from Map",
                        testing.IsEqual(np.array(o1), x + y).all() and testing.IsEqual(np.array(o2), x - y).all())

    compiled = map.Compile("host", "test", "predict")
    compiled.ComputeMultiple([i1, i2], [o1, o2])
    testing.ProcessTest("Testing multiple inputs and outputs from CompiledMap",
                        testing.IsEqual(np.array(o1), x + y).all() and testing.IsEqual(np.array(o2), x - y).all())

```
See also [CallbackDesign](https://intelligentdevices.visualstudio.com/EMLL/_git/ELL/pullrequest/1264?_a=files&path=%2Finterfaces%2Fcommon%2Fdoc%2FCallbackDesign.md)

Related work items: #2190

+++ libraries/model/include/Submodel.h
@@ -68,10 +68,10 @@ namespace model
         Model& GetModel() { return _model; }
 
         /// <summary> Returns the number of input ports for this submodel </summary>
-        int NumInputs() const { return _inputs.size(); }
+        int NumInputs() const { return static_cast<int>(_inputs.size()); }
 
         /// <summary> Returns the number of output ports for this submodel </summary>
-        int NumOutputs() const { return _outputs.size(); }
+        int NumOutputs() const { return static_cast<int>(_outputs.size()); }
 
         /// <summary> Returns the input ports for this submodel </summary>
         const std::vector<const InputPortBase*>& GetInputs() const { return _inputs; }
